% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVDJ.CellType.Connection.R
\name{runVDJ.CellType.Connection}
\alias{runVDJ.CellType.Connection}
\title{runVDJ.CellType.Connection}
\usage{
runVDJ.CellType.Connection(
  object,
  cell.type.label,
  clone.id = "clone_id",
  cell.type = "all"
)
}
\arguments{
\item{object}{The Seurat S4 object}

\item{cell.type.label}{The cell type annotation labels in Seurat S4 object}

\item{clone.id}{Default "clone_id"; The result of changeo has been intergrated in Seurat S4 object. The clonetypes information have been save in "clone_id" by default.}
}
\value{
Heatmap of clone sharing between cell types
}
\description{
The runVDJ.CellType.Connection function will caculate the clonetype connection between different cell types.
}
\details{
The description of this runVDJ.CellType.Connection
}
