% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVDJ.CloneFreqLevels.R
\name{runVDJ.CloneFreqLevels}
\alias{runVDJ.CloneFreqLevels}
\title{runVDJ.CloneFreqLevels}
\usage{
runVDJ.CloneFreqLevels(object, clone.id = "clone_id")
}
\arguments{
\item{object}{The Seurat S4 object}

\item{clone.id}{Default "clone_id"; The result of change-o has been integrated in Seurat S4 object. The clonetypes information have been save in "clone_id" by default.}
}
\value{
Seurat S4 object with CloneFreqLevels labels saved
}
\description{
The runVDJ.CloneFreqLevels function will calculate clonetype frequency levels.
}
\details{
The description of this runVDJ.CloneFreqLevels
}
