% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVDJ.CloneSharing.R
\name{runVDJ.CloneSharing}
\alias{runVDJ.CloneSharing}
\title{runVDJ.CloneSharing}
\usage{
runVDJ.CloneSharing(object, cell.type.label, clone.id)
}
\arguments{
\item{object}{The Seurat S4 object}

\item{cell.type.label}{The cell type annotation labels in Seurat S4 object}

\item{clone.id}{Default "clone_id"; The result of change-o has been integrated in Seurat S4 object. The clonetypes information have been save in "clone_id" by default.}

\item{cell.type}{The selected cell type in annotation labels will be used to calculate the clonetype connection with other cell types.}
}
\value{
Heatmap of clone sharing between cell types
}
\description{
The runVDJ.CloneSharing function will calculate the clonetype sharing between different cell types.
}
\details{
The description of this runVDJ.CloneSharing
}
