% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVDJ.diffUsage.R
\name{runVDJ.diffUsage}
\alias{runVDJ.diffUsage}
\title{runVDJ.diffUsage}
\usage{
runVDJ.diffUsage(
  object,
  v.call = "germline_v_call",
  sample.label = "samples",
  group.label = "group",
  GroupA,
  GroupB,
  reference,
  min.cells = 10
)
}
\arguments{
\item{object}{The Seurat S4 object}

\item{v.call}{Default "germline_v_call"; The result of change-o has been integrated in Seurat S4 object. The V region information have been save in "germine_v_call" by default}

\item{sample.label}{The sample labels in Seurat S4 object.}

\item{group.label}{The group labels in Seurat S4 object.}

\item{GroupA}{Vector of samples names belonging to group A}

\item{GroupB}{Vector of samples names belonging to group B}

\item{reference}{Vector of samples names belonging to reference}

\item{min.cells}{Default 10; The effective sample contains min cell numbers}
}
\value{
Complex dotplot of 3 groups' IGHV Usage
}
\description{
The runVDJ.diffUsage function will calculate the group specific IGHV Usage.
}
\details{
The description of this runVDJ.diffUsage
}
