% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVDJ.Diversity.R
\name{runVDJ.Diversity}
\alias{runVDJ.Diversity}
\title{runVDJ.Diversity}
\usage{
runVDJ.Diversity(object, cell.type.label, clone.id = "clone_id")
}
\arguments{
\item{object}{The Seurat S4 object}

\item{cell.type.label}{The cell type annotation labels in Seurat S4 object}

\item{clone.id}{Default "clone_id"; The result of change-o has been integrated in Seurat S4 object. The clonetypes information have been save in "clone_id" by default.}
}
\value{
Bar plot of clonetype diversity
}
\description{
The runVDJ.Diversity function will calculate the clonetype diversity.
}
\details{
The description of this runVDJ.Diversity
}
