% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVDJ.intergrated.R
\name{runVDJ.intergrated}
\alias{runVDJ.intergrated}
\title{runVDJ.intergrated}
\usage{
runVDJ.intergrated(tsv.file, Seurat.object, runSHM = T)
}
\arguments{
\item{tsv.file}{The file paths of change-o result. such as "~/*/filtered_heavy_germ-pass.tsv"}

\item{runSHM}{select T/F; Determined to calculate the mutation frequency}

\item{object}{The Seurat S4 object}
}
\value{
Integrated S4 object
}
\description{
The runVDJ.intergrated function will integrated the result of change-o with the Seurat.S4 object.
}
\details{
The description of this runVDJ.intergrated.
}
