% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVDJ.Selec.Strength.R
\name{runVDJ.Selec.Strength}
\alias{runVDJ.Selec.Strength}
\title{runVDJ.Selec.Strength}
\usage{
runVDJ.Selec.Strength(
  object,
  group.label = "group",
  isotype.label,
  isotype = "IGHM",
  cell.type.label,
  cell.type,
  clone.id = "clone_id"
)
}
\arguments{
\item{object}{The Seurat S4 object}

\item{group.label}{The group labels in Seurat S4 object}

\item{isotype.label}{Default "c_call"; The result of change-o has been integrated in Seurat S4 object. The isotype information have been save in "c_call" by default}

\item{cell.type.label}{The cell type annotation labels in Seurat S4 object}

\item{cell.type}{The selected cell type in annotation labels will be used to calculate the selection strength}

\item{clone.id}{Default "clone_id"; The result of change-o has been integrated in Seurat S4 object. The clonetypes information have been save in "clone_id" by default.}
}
\value{
Complex plot selection strength in cdr3 and fwr
}
\description{
The runVDJ.Selec.Strength function will caculate the selection strength in each group.
}
\details{
The description of this runVDJ.Selec.Strength
}
