% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVDJ.Usage.R
\name{runVDJ.Usage}
\alias{runVDJ.Usage}
\title{runVDJ.Usage}
\usage{
runVDJ.Usage(
  object,
  v.call = "germline_v_call",
  sample.label = "samples",
  group.label = "group",
  min.cells = 10
)
}
\arguments{
\item{object}{The Seurat S4 object}

\item{v.call}{Default "germline_v_call"; The result of change-o has been integrated in Seurat S4 object. The V region information have been save in "germine_v_call" by default.}

\item{sample.label}{The sample labels in Seurat S4 object.}

\item{group.label}{The group labels in Seurat S4 object.}

\item{min.cells}{Default 10; The effective sample contains min cell numbers.}
}
\value{
IGHV Usage box plot
}
\description{
The runVDJ.Usage function will calculate the IGHV fragment Usage.
}
\details{
The description of this runVDJ.Usage.
}
